move introspection config factory into introspection datasource package and export functions
